fixed comentarios PR
