--- src/main/java/org/drg/App.java
@@ -1,20 +1,0 @@
-package org.drg;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class App
-{
-    public static void main( String[] args )
-    {
-        SpringApplication.run(App.class, args);
-    }
-
- /*   @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }*/
-
-
-}
