--- src/main/java/org/drg/dtos/ErrorDto.java
@@ -1,12 +1,0 @@
-package org.drg.dtos;
-
-import lombok.Data;
-
-@Data
-public class ErrorDto {
-
-    private String code;
-    private String message;
-    private String detail;
-
-}
