--- src/main/java/org/drg/dtos/ErrorsDto.java
@@ -1,13 +1,0 @@
-package org.drg.dtos;
-
-import lombok.Data;
-import org.drg.dtos.ErrorDto;
-
-import java.util.List;
-
-@Data
-public class ErrorsDto {
-
-    private List<ErrorDto> errorDtoList;
-
-}
