--- src/main/java/org/drg/dtos/requests/UserRequestDto.java
@@ -1,39 +1,0 @@
-package org.drg.dtos.requests;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Data;
-import org.springframework.validation.annotation.Validated;
-
-@Validated
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-11-02T19:44:59.902Z[GMT]")
-
-@Builder
-@Data
-public class UserRequestDto {
-
-    @JsonProperty("name")
-    private String name;
-
-    @JsonProperty("surname")
-    private String surname;
-
-    @JsonProperty("username")
-    private String username;
-
-    @JsonProperty("mail")
-    private String mail;
-    
-    @JsonProperty("password")
-    private String password;
-
-    @JsonProperty("address")
-    private String address;
-
-    @JsonProperty("floor")
-    private String floor;
-
-    @JsonProperty("door")
-    private String door;
-
-}
