--- src/main/java/org/drg/exceptions/DataTypeErrorException.java
@@ -1,9 +1,0 @@
-package org.drg.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class DataTypeErrorException extends RuntimeException{
-
-}
