--- src/main/java/org/drg/models/UserEntity.java
@@ -1,43 +1,0 @@
-package org.drg.models;
-
-
-import lombok.Data;
-
-import javax.persistence.*;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Entity (name = "t_user")
-@Data
-public class UserEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @NotBlank(message = "Name is mandatory")
-    private String name;
-    
-    @NotBlank(message = "Surname is mandatory")
-    private String surname;
-    
-    @NotBlank(message = "Username is mandatory")
-    private String username;
-
-    @NotBlank(message = "Mail is mandatory")
-    @Email(message = "Email is not valid", regexp = "(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])")
-    private String mail;
-
-    @NotBlank(message = "Password is mandatory")
-    private String password;
-
-    @NotBlank(message = "Address is mandatory")
-    private String address;
-
-    @NotBlank(message = "Floor is mandatory")
-    private String floor;
-
-    @NotBlank(message = "Door is mandatory")
-    private String door;
-
-}
