--- src/main/java/org/drg/security/SecurityConfiguration.java
@@ -1,9 +1,0 @@
-package org.drg.security;
-
-
-public class SecurityConfiguration {
-
-
-
-
-}
