+++ src/main/java/org/drg/springcourt/App.java
@@ -1,0 +1,29 @@
+package org.drg.springcourt;
+
+import org.dozer.DozerBeanMapper;
+import org.dozer.Mapper;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+import java.util.Arrays;
+
+@SpringBootApplication
+public class App
+{
+    public static void main( String[] args )
+    {
+        SpringApplication.run(App.class, args);
+    }
+
+ /*   @Bean
+    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
+    }*/
+
+    @Bean
+    public Mapper mapper() {
+        return new DozerBeanMapper(Arrays.asList("dozer_mapping.xml"));
+    }
+
+}
