+++ src/main/java/org/drg/springcourt/dtos/ErrorDto.java
@@ -1,0 +1,12 @@
+package org.drg.springcourt.dtos;
+
+import lombok.Data;
+
+@Data
+public class ErrorDto {
+
+    private String code;
+    private String message;
+    private String detail;
+
+}
