+++ src/main/java/org/drg/springcourt/dtos/ErrorListDto.java
@@ -1,0 +1,12 @@
+package org.drg.springcourt.dtos;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class ErrorListDto {
+
+    private List<ErrorDto> list;
+
+}
