+++ src/main/java/org/drg/springcourt/dtos/requests/UserRequestDto.java
@@ -1,0 +1,34 @@
+package org.drg.springcourt.dtos.requests;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Data;
+import org.springframework.validation.annotation.Validated;
+
+@Validated
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-11-02T19:44:59.902Z[GMT]")
+@Builder
+@Data
+public class UserRequestDto {
+
+    private Long id;
+
+    @JsonProperty("name")
+    private String name;
+
+    @JsonProperty("surname")
+    private String surname;
+
+    @JsonProperty("username")
+    private String username;
+
+    @JsonProperty("mail")
+    private String mail;
+    
+    @JsonProperty("password")
+    private String password;
+
+/*    @JsonProperty("address")
+    private List<AddressRequestDto> addressList;*/
+
+}
