+++ src/main/java/org/drg/springcourt/dtos/responses/AddressResponseDto.java
@@ -1,0 +1,20 @@
+package org.drg.springcourt.dtos.responses;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class AddressResponseDto{
+
+    @JsonProperty("address")
+    private String address;
+
+    @JsonProperty("floor")
+    private String floor;
+
+    @JsonProperty("door")
+    private String door;
+
+}
