+++ src/main/java/org/drg/springcourt/exceptions/DataTypeErrorException.java
@@ -1,0 +1,9 @@
+package org.drg.springcourt.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class DataTypeErrorException extends RuntimeException{
+
+}
