+++ src/main/java/org/drg/springcourt/repositories/AddressRepository.java
@@ -1,0 +1,10 @@
+package org.drg.springcourt.repositories;
+
+import org.drg.springcourt.models.AddressEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AddressRepository extends JpaRepository<AddressEntity, Long> {
+
+}
