+++ src/main/java/org/drg/springcourt/repositories/UserRepository.java
@@ -1,0 +1,10 @@
+package org.drg.springcourt.repositories;
+
+import org.drg.springcourt.models.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    Boolean existsByUsername(String username);
+}
