+++ src/main/java/org/drg/springcourt/security/SecurityConfiguration.java
@@ -1,0 +1,9 @@
+package org.drg.springcourt.security;
+
+
+public class SecurityConfiguration {
+
+
+
+
+}
