+++ src/main/java/org/drg/springcourt/services/AddressService.java
@@ -1,0 +1,17 @@
+package org.drg.springcourt.services;
+
+import org.drg.springcourt.dtos.requests.UserRequestDto;
+import org.drg.springcourt.dtos.responses.UserResponseDto;
+import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+
+@Service
+public interface AddressService {
+
+    Collection<UserResponseDto> findAll();
+
+    UserResponseDto saveUser(UserRequestDto user);
+
+    Collection<UserResponseDto> findAllUsersSameStreet(UserRequestDto userRequestStreetDto);
+}
