+++ src/main/resources/data.sql
@@ -1,9 +1,14 @@
 
 
 -- saving users
-INSERT INTO t_user (id,name,surname,username,mail,password,address,floor,door)
-    VALUES(99, 'David','Rey','dreyg','d.reyg30@gmail.com','1234','calle las adelfas 2','4','D');
-INSERT INTO t_user (id,name,surname,username,mail,password,address,floor,door)
-    VALUES(100, 'Antonio','Arellano','aarellanom','a.arellano@gmail.com','1234','calle petroleo 3','5','A');
+INSERT INTO users (id,name,surname,username,mail,password)
+    VALUES(99, 'David','Rey','dreyg','d.reyg30@gmail.com','1234');
+INSERT INTO users (id,name,surname,username,mail,password)
+    VALUES(100, 'Antonio','Arellano','aarellanom','a.arellano@gmail.com','1234');
 
 
+INSERT INTO address (id,address,door,floor,user_id)
+    VALUES(50, 'Calle Adelfas 2C','4','D',99);
+
+INSERT INTO address (id,address,door,floor,user_id)
+    VALUES(51, 'Calle Petroleo 7','5','A',100);
